Move Middle & TitleInput from project/common to project/edit
