+++ frontend/src/components/project/edit/ProjectEdit.jsx
@@ -8,8 +8,8 @@ import Color from "@components/project/Creates/Color"
 import Privacy from "@components/project/Creates/Privacy"
 import Type from "@components/project/Creates/Type"
 import { getProjectColor } from "@components/project/Creates/palettes"
-import Middle from "@components/project/common/Middle"
-import TitleInput from "@components/project/common/TitleInput"
+import Middle from "@components/project/edit/Middle"
+import TitleInput from "@components/project/edit/TitleInput"
 import EditBox from "@components/project/edit/EditBox"
 
 import { patchProject } from "@api/projects.api"
